refactor: extract upgrade filter preview modal
